songs-microservice commit

+++ resource-service/src/main/java/com/goropai/resourceservice/controller/ResourceController.java
@@ -1,16 +1,24 @@
 package com.goropai.resourceservice.controller;
 
 import com.goropai.resourceservice.entity.Mp3File;
+import com.goropai.resourceservice.entity.dto.ResourceIdResponse;
+import com.goropai.resourceservice.entity.dto.ResourceIdsResponse;
 import com.goropai.resourceservice.service.MetadataService;
 import com.goropai.resourceservice.service.ResourceService;
+import com.goropai.songservice.entity.Mp3Metadata;
 import jakarta.persistence.PersistenceException;
 import jakarta.transaction.Transactional;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
+import org.springframework.web.server.ResponseStatusException;
+import reactor.core.publisher.Mono;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -36,24 +44,39 @@ public class ResourceController {
      * 500 Internal Server Error – An error occurred on the server.
      */
     @PostMapping(path = "/resources", consumes = "audio/mpeg")
-    public ResponseEntity<Integer> uploadAudio(@RequestBody byte[] audioData) {
-        Mp3File savedFile;
-        try {
-            savedFile = resourceService.save(audioData);
-        }
-        catch (PersistenceException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-
-        try {
-            /*metadataService.parseAndSave(savedFile.getId(), savedFile.getData())
-                    .subscribe();*/
-        }
-        catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-
-        return ResponseEntity.status(HttpStatus.OK).body(savedFile.getId());
+    @Transactional
+    public Mono<ResponseEntity<ResourceIdResponse>> uploadAudio(@RequestBody byte[] audioData) {
+        return Mono.fromCallable(() -> resourceService.save(audioData))
+                .onErrorResume(PersistenceException.class, ex -> Mono.error(new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR)))
+                .flatMap(savedFile ->
+                        {
+                            try {
+                                return metadataService.parseAndSave(savedFile.getId(), savedFile.getData())
+                                        .onErrorResume(
+                                                ex -> {
+                                                    try {
+                                                        resourceService.deleteById(savedFile.getId());
+                                                    }
+                                                    catch (Exception e) {
+                                                        return Mono.error(new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR));
+                                                    }
+                                                    if (ex instanceof WebClientResponseException) {
+                                                        return Mono.error(ex);
+                                                    }
+                                                    else {
+                                                        return Mono.error(new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR));
+                                                    }
+                                                }
+                                        );
+                            } catch (IOException e) {
+                                return Mono.error(new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR));
+                            }
+                        }
+                )
+                .map(success -> ResponseEntity.status(HttpStatus.OK).body(
+                        new ResourceIdResponse(Optional.ofNullable(success.getBody())
+                                .map(Mp3Metadata::getId).orElse(null))))
+                .onErrorResume(ResponseStatusException.class, ex -> Mono.error(new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR)));
     }
 
     /**
@@ -66,12 +89,16 @@ public class ResourceController {
      * 404 Not Found – Resource with the specified ID does not exist.
      * 500 Internal Server Error – An error occurred on the server.
      */
-    @GetMapping(path = "/resources/{id}", produces = "application/json")
-    public ResponseEntity<Mp3File> getAudio(@PathVariable @Validated int id) {
+    @GetMapping(path = "/resources/{id}")
+    public ResponseEntity<byte[]> getAudio(@PathVariable @Validated int id) {
         try {
             Optional<Mp3File> found = resourceService.getById(id);
-            return found.map(m -> ResponseEntity.status(HttpStatus.OK).body(m))
-                    .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+            return found.map(m -> ResponseEntity.status(HttpStatus.OK)
+                            .contentType(MediaType.parseMediaType("audio/mpeg"))
+                            .body(m.getData()))
+                    .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .body("{\"404\": \"Resource not found\"}".getBytes()));
         }
         catch (PersistenceException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
@@ -95,7 +122,7 @@ public class ResourceController {
      */
     @DeleteMapping(path = "/resources")
     @Transactional
-    public ResponseEntity<List<Integer>> deleteAudio(@RequestParam(name = "id") String ids) {
+    public ResponseEntity<ResourceIdsResponse> deleteAudio(@RequestParam(name = "id") String ids) {
         if (ids.length() <= 0 || ids.length() > 200) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
@@ -103,12 +130,16 @@ public class ResourceController {
                 .map(Integer::valueOf)
                 .filter(resourceService::existsById)
                 .toList();
+        if (correctIds.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ResourceIdsResponse());
+        }
         try {
             correctIds.forEach(resourceService::deleteById);
+            metadataService.deleteMetadata(correctIds).subscribe();
         }
         catch (PersistenceException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
-        return ResponseEntity.status(HttpStatus.OK).body(correctIds);
+        return ResponseEntity.status(HttpStatus.OK).body(new ResourceIdsResponse(correctIds));
     }
 }
